# api/openapi.yaml
openapi: 3.1.0
info:
  title: Island Media Co connected assessment API
  version: 2.0.0
  description: >-
    Fictional mock boundary for the Frameworks technical assessment. Protected
    prototype endpoints use X-Prototype-User-Id to represent the selected
    seeded or newly registered user. Production authentication is outside scope.
servers:
  - url: /api
paths:
...
  /booking-requests:
    post:
      summary: Submit a non-binding advertising request
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequestCreate"
      responses:
        "201":
          description: New request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingRequest"
        "200":
          description: Existing request returned for a repeated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /management/booking-requests:
    get:
      summary: List requests for management review
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - in: query
          name: status
          schema:
            type: string
            enum: [submitted, information_required, approved, declined]
      responses:
        "200":
          description: Request list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookingRequestSummary"
        "403":
          $ref: "#/components/responses/Forbidden"
  /management/booking-requests/{requestId}:
    get:
      summary: Get a request with client and current availability
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Management request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingRequestDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Apply a management decision to a request
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementDecision"
      responses:
        "200":
          description: Updated request, optionally with draft contract ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingRequestDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /management/contracts/{contractId}/issue:
    post:
      summary: Issue a draft contract to its client
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - in: path
          name: contractId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contract issued and client-visible event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
    PrototypeUserId:
      in: header
      name: X-Prototype-User-Id
      required: true
      schema:
        type: string
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
  schemas:
...
    Organisation:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        contractCount:
          type: integer
          minimum: 0
...
    IndicativeRate:
      type: object
      required: [currency, label]
      properties:
        currency:
          type: string
          const: GBP
        amount:
          type: [number, "null"]
        unit:
          type: [string, "null"]
        monthlyEquivalent:
          type: [number, "null"]
        label:
          type: string
    AvailabilitySummary:
      type: object
      required: [state, reason, calculatedAt]
      properties:
        state:
          type: string
          enum: [available, unavailable, confirmation_required]
        reason:
          type: string
        calculatedAt:
          type: string
          format: date-time
        availableAssetCount:
          type: [integer, "null"]
        availableCapacity:
          type: [integer, "null"]
        totalCapacity:
          type: [integer, "null"]
        freshestVerificationAt:
          type: [string, "null"]
          format: date-time
    ProductSearchResult:
      type: object
      required:
        [
          id,
          name,
          mediaOwnerName,
          mediaType,
          locationNames,
          allocationModel,
          indicativeRate,
          minimumTermDays,
          availability,
        ]
      properties:
        id:
          type: string
        name:
          type: string
        mediaOwnerName:
          type: string
        mediaType:
          type: string
        locationNames:
          type: array
          items:
            type: string
        allocationModel:
          type: string
          enum: [exclusive_asset, capacity_pool]
        indicativeRate:
          $ref: "#/components/schemas/IndicativeRate"
        minimumTermDays:
          type: integer
        availability:
          $ref: "#/components/schemas/AvailabilitySummary"
...
    BookingRequestCreate:
      type: object
      required: [productId, startDate, endDate, budget, objective]
      properties:
        productId:
          type: string
        requestedAssetId:
          type: [string, "null"]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
          minimum: 0
        objective:
          type: string
          minLength: 5
        notes:
          type: [string, "null"]
    HistoryEntry:
      type: object
      required: [at, actor, action]
      properties:
        at:
          type: string
          format: date-time
        actor:
          type: string
        action:
          type: string
        note:
          type: [string, "null"]
    BookingRequest:
      allOf:
        - $ref: "#/components/schemas/BookingRequestCreate"
        - type: object
          required: [id, organisationId, status, createdAt, history]
          properties:
            id:
              type: string
            organisationId:
              type: string
            status:
              type: string
              enum: [submitted, information_required, approved, declined]
            createdAt:
              type: string
              format: date-time
            draftContractId:
              type: [string, "null"]
            history:
              type: array
              items:
                $ref: "#/components/schemas/HistoryEntry"
    BookingRequestSummary:
      type: object
      required:
        [id, organisationName, productName, startDate, endDate, status, createdAt]
      properties:
        id:
          type: string
        organisationName:
          type: string
        productName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        attentionReason:
          type: [string, "null"]
    BookingRequestDetail:
      allOf:
        - $ref: "#/components/schemas/BookingRequest"
        - type: object
          required: [organisation, product, currentAvailability]
          properties:
            organisation:
              $ref: "#/components/schemas/Organisation"
            product:
              $ref: "#/components/schemas/ProductSearchResult"
            currentAvailability:
              $ref: "#/components/schemas/AvailabilitySummary"
    ManagementDecision:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [request_information, approve, decline]
        note:
          type: [string, "null"]
        selectedAssetId:
          type: [string, "null"]
    ContractItem:
      type: object
      required: [id, productId, quantity, lineTotal]
      properties:
        id:
          type: string
        productId:
          type: string
        assetId:
          type: [string, "null"]
        capacityPoolId:
          type: [string, "null"]
        quantity:
          type: integer
        unitRate:
          type: [number, "null"]
        rateUnit:
          type: [string, "null"]
        lineTotal:
          type: number
...
    Contract:
      type: object
      required:
        [
          id,
          organisationId,
          status,
          version,
          startDate,
          endDate,
          currency,
          total,
          items,
          history,
        ]
      properties:
        id:
          type: string
        organisationId:
          type: string
        bookingRequestId:
          type: [string, "null"]
        status:
          type: string
          enum:
            [draft, issued, change_requested, accepted, active, completed, cancelled]
        version:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        currency:
          type: string
          const: GBP
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContractItem"
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: [object, "null"]
          additionalProperties: true
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Record not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The selected prototype user cannot access this record or action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: FORBIDDEN
            message: You cannot access this organisation's record.
    Conflict:
      description: State or availability changed before the action completed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INVENTORY_CONFLICT
            message: No suitable inventory remains for the requested dates.
    ServiceUnavailable:
      description: Simulated or real temporary service failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: SERVICE_UNAVAILABLE
            message: Please try again. The service is temporarily unavailable.